+++ libs/rxjs/array/src/lib/filter-intersects.ts
@@ -6,11 +6,11 @@ import { isObservable, Observable, ObservableInput, of, OperatorFunction } from 
 import { map, switchMap } from 'rxjs/operators';
 
 /**
- * Returns an Observable Array containing unique values that are in both the source and provided input Array or Set
+ * Returns an Observable Array containing filtered values that are in both the source in the provided input Array or Set
  *
- * @category Compare
+ * @category Filter
  *
- * @see [[filterIntersects]] operator for an Array containing potential duplicate intersections
+ * @see [[intersects]] operator for an Array of unique intersection items
  *
  * @typeParam T Item type contained in the Array/Set
  *
@@ -20,32 +20,31 @@ import { map, switchMap } from 'rxjs/operators';
  * Returns the intersection between the source array and the passed static array
  * ```ts
  * const input = ['a', 'b', 'd', 'a', 'b'];
- * of(input).pipe(intersects(['a', 'd'])).subscribe();
+ * of(input).pipe(filterIntersects(['a', 'd'])).subscribe();
  * ```
- * Output: `'a', 'd'`
+ * Output: `'a', 'd', 'a'`
  *
  * @example
  * Returns the intersection between the source array and the passed Observable array
  * ```ts
  * const input = ['a', 'b', 'd', 'a', 'b'];
- * of(input).pipe(intersects(of(['a', 'd']))).subscribe();
+ * of(input).pipe(filterIntersects(of(['a', 'd']))).subscribe();
  * ```
- * Output: `'a', 'd'`
+ * Output: `'a', 'd', 'a'`
  *
- * @returns An Observable that emits an array of the intersection of input and source arrays.
+ * @returns An Observable that emits an Array of the intersection of input and source arrays.
  */
-export function intersects<T extends unknown>(
+export function filterIntersects<T extends unknown>(
   input: T[] | Set<T> | ObservableInput<T[] | Set<T>>,
 ): OperatorFunction<T[] | Set<T>, T[]> {
   return (source) =>
     ((isObservable(input) ? input : of(input)) as Observable<T[] | Set<T>>).pipe(
-      map((val) => new Set(val)),
-      switchMap((inputValue: Set<T>) =>
-        source.pipe(
-          map((value) => {
-            return [...new Set<T>(value)].filter((x) => inputValue.has(x));
-          }),
-        ),
-      ),
+      map((value) => new Set(value)),
+      switchMap((inputValue) => {
+        return source.pipe(
+          map((value) => [...value]),
+          map((value) => value.filter((v) => inputValue.has(v))),
+        );
+      }),
     );
 }