+++ libs/rxjs/array/CHANGELOG.md
@@ -7,6 +7,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+This release contains a few breaking changes and renaming of operators
+
 ### Added
 
 - `isSubsetOf` operator that takes an Array or Set and checks if the source Array or Set is a subset of it
@@ -15,6 +17,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Changed
 
 - All operators now accept Set where they also currently accept an Array, either as a source or as an input value
+- `difference` and `intersection` no longer accept a mapping method and only accepts an `Array` or `Set` value to return the pure difference
+- `differenceWith` has been renamed `filterDifference` and `intersectionWith` is now `filterIntersection` - both no longer accepts a predicate method.
 
 ## [3.1.0] - 2021-01-14
 
@@ -48,7 +52,7 @@ coverage that provided various bug fixes.
 
 ### Changed
 
-- `difference`, `differenceWith`, `intersection` and `intersectionWith` now use `concatMap` instead of `switchMap`
+- `difference`, `filterDifference`, `intersection` and `intersectionWith` now use `concatMap` instead of `switchMap`
 - `binarySearch` improvements, now returns the unsorted and sorted array
 - Functions that return only truthy values (e.g. `filter*`, `find`, `every`, etc) with predicates treat `0` as truthy
   but `''` as falsy
@@ -121,6 +125,6 @@ coverage that provided various bug fixes.
 
 - `binarySearch` operator that provides a way to search for a value within an Observable array
 - `difference` operator that returns the difference between two arrays with optional modifier method
-- `differenceWith` operator that returns the difference between two arrays with optional predicate method
+- `filterDifference` operator that returns the difference between two arrays with optional predicate method
 - `intersects` operator that returns the intersection between two arrays with optional modifier method
-- `intersectsWith` operator that returns the intersection between two arrays with optional predicate method
+- `filterIntersects` operator that returns the intersection between two arrays with optional predicate method